backend/rag/requirements.txt
# RAG Pipeline Dependencies
# Note: Main dependencies are managed in backend/pyproject.toml
# This file is for reference and manual installation only

# Core LlamaIndex
llama-index-core>=0.10.0
llama-index>=0.10.0

# Embeddings
llama-index-embeddings-openai>=0.1.0

# Vector stores
llama-index-vector-stores-pinecone>=0.1.0
pinecone>=5.0.0

# Other dependencies
python-dotenv>=1.0.0
openai>=1.0.0
